+++ media/src/main/kotlin/com/imashnake/animite/media/MediaMediumGrid.kt
@@ -12,7 +12,6 @@ import androidx.compose.ui.unit.dp
 import com.imashnake.animite.api.anilist.sanitize.media.Media
 import com.imashnake.animite.core.extensions.plus
 import com.imashnake.animite.core.ui.LocalPaddings
-import com.imashnake.animite.core.ui.MediaCard
 import kotlinx.collections.immutable.ImmutableList
 
 @Composable
